buttons: Connect PopUp Confirmation View for topic/streams via keypress.
On pressing `MARK_STREAM_AS_READ` or `MARK_TOPIC_AS_READ` from
`StreamButton` or `TopicButton` respectively, a confirmation display pops up
corresponding to the stream or topic asking for user's confirmation for
marking the stream or topic as read.

Event handling for update of message flags already takes care of UI update
on recieving the response for this event.

Tests to be added.